add IRQ support + syncbypass
This attempts to add in IRQ support. The main idea is to create
a "scoreboard" which tracks interrupt outputs from the four machines
and reduces them to a single vector of shared interrupt state.

Because the IRQ/WAIT instructions can also clear as well as set
a bit, a "strobe" is added to the interrupt function of the machine.
When the "strobe" is set, it informs the scoreboard that a given
bit of the interrupt shared state is desired to be updated.

From there, every cycle we evaluate if a bit should be updated,
with strobes from the highest numbered machines taking priority
over lower numbered machines. This is finally reduced to a single
8-bit vector that can be used by all the machines.

This also adds a "sync_bypass" option (and a synchronizer on
the GPIOs), as well as breaks out some more internal state for debug
feedback via an upper-level register harness.