rework osr/isr auto-pull priorities
The mystery of how they solve the critical path has been solved.
Instead of computing the shift and all in a single cycle, what
happens is you need a "look-ahead" count that is used to trigger
a pull in the case that the element you've pulled empties the
OSR. There is also a provision for the pull to stall the OUT
instruction. Basically, a small amount of pipelining is done
at the expense of creating some ambiguity in the MOV+autofill case.