fix OUT behavior
1) OUT always executes in not auto-pull mode (it should be non-blocking).
The spec is a little vague on this, but it doesn't say anywhere that
OUT should stall (in fact it says you need to use a stalling PULL
explicitly if you don't have auto-pull on). So I think OUT without
autopull on should always execute immediately, even if the OUT count
has been exceeded.

The blocking behavior is implied in the case that auto-pull is turned
on, since it's meant to be a fusion of blocking PULL + OUT in that case...

2) Add NULL shift target: still do the shift, even if the
destination is NULL!