rewrite sideset decoder so that the form is explicit
The shift left/shift right thing is probably insanely clever
but it also relies upon the synthesis tool obeying bit width
and operator precedent to the spec. I also suspect there was
a bug having to do with bit widths and the optional set bit,
this form seems to have cleared that up.