refactor ISR into a more legible form
The original form uses some...clever rotate-to-shift thing.
I think it's correct but also the spec doesn't call for a rotation
of upper bits to lower bits, and it's hard to read.

This form is a little clearer (at least to me) and it compiles
down to slightly less gates because it's not doing a rotate
that's unneeded, just a shift.