remove redundant wiring that confuses synthesis and simulation
for some reason the original author wired all the busses together
within the device, using tasks to push things onto it and pull
them off. This has caused a lot of problems in correctness; in
this case, I get different results on verilator versus xsim.

Wiring the ISR directly to the FIFO instead of sharing it with
all the other potential dout consequences eliminates this ambiguity
and now both simulators agree.