remove shared bus resource, add fast CDC mode for DMA
DMA cycles need to drop their request within 2 cycles of
a write request on APB otherwise it will re-trigger the DMA request:

T0  T1  T2  T3
--------------
W   I   I   S

T0 is the last "write" on the APB of the DMA data.

T1 and T2 are internal cycles for the DMAC as it cleans up
the transfer.

T3 is when it samples the level-sensitive request signal again.

Thus, the APB would assert the write on T0, but the PIO block
won't sample it until T1. It must drop the request by the end
of T2, so that on the edge T3 it is not high again.

This is a very tight timing when PCLK:PIO_CLK is equal to 1:1.
In order to meet this timing, a CDC mode bit is provided in
the supplemental configuration area, where each channel can have
its push/pull pulse signals either run through a regular CDC,
or it can just go through a simple edge-to-pulse converter.

For ratios slower than 1:4 bus:pio clock, the regular CDC
can meet timing, and thus any ratio less than 1:4 should
be acceptable (1:5, 1:6, etc..)

For ratios 1:2 and 1:1, the pulse converter should be used,
and it is mandatory that the edges are synchronous and aligned
between these two domains as there is effectively no CDC
present.

A ratio of 1:3 or any other fractional ratios are not compatible
with DMA. However, any ratio can be used if DMA is not required.

This commit also includes some cleanup of the shared bus
motif that was used in the original code. I think separating out
the busses makes the code clearer and reduces the risk of cross-task
side effects that are not obvious, at the expense of slightly
higher resource utilization.