fix OUT EXEC running an OUT PC or MOV PC
need to look into whether we should also be allowing JMPs
to execute during an EXEC context

my read of the spec is that an OUT EXEC would insert an instruction
into the stream and then continue on to the next instruction

but....the spec is a little ambiguous if that means that a JMP
instruction in the stream should be "executed" and then you resume
on your merry way executing the next instruction after the OUT
(ie. it's a "subroutine" style behavior) or if it is meant to
permanently side-effect the PC as the OUT PC or MOV PC instructions
are explicitly designated to do.