fix exec of jmp-on-stalled exec, and make irq_stb's pulsed not edge triggered
- exec of jmp on stalled exec requires a precise stall signal. the
previous implementation relied on the coding of the instruction to
determine when an exec could be allowed, but when you also allow
jumps, it turns out that that the coding of the jump opcode (000)
is identical to the coding of ... no instruction to execute.
so we can't rely on that. this adds a `waiting1` signal that indicates
that the decode was valid, which we can use as a gate for the exec-exec
case.

- make irq_stb level based, not edge based. Why? because with
clkdiv = 1.0, successive irq_stb's will be missed.

there is now a test case for clkdiv=1.0 in the test bench to
catch this.